// src/common/enums/error-code.enum.ts
export enum ErrorCode {
  //User
  INVALID_PASSWORD = 'INVALID_PASSWORD',
  USER_NOT_FOUND = 'USER_NOT_FOUND',
  USER_UNAUTHORIZED = 'USER_UNAUTHORIZED',
  USER_FORBIDDEN = 'USER_FORBIDDEN',
  PHONE_NUMBER_ALREADY_EXISTS = 'PHONE_NUMBER_ALREADY_EXISTS',
  EMAIL_ALREADY_EXISTS = 'EMAIL_ALREADY_EXISTS',
  INVALID_VERIFICATION_CODE = 'INVALID_VERIFICATION_CODE',
  VERIFICATION_CODE_EXPIRED = 'VERIFICATION_CODE_EXPIRED',
  RESET_TOKEN_MISSING = 'RESET_TOKEN_MISSING',
  INVALID_RESET_TOKEN = 'INVALID_RESET_TOKEN',
  INVALID_RESET_TOKEN_PAYLOAD = 'INVALID_RESET_TOKEN_PAYLOAD',

  //Post
  POST_NOT_FOUND = 'POST_NOT_FOUND',
  POST_NOT_APPROVED = 'POST_NOT_APPROVED',
  POST_ALREADY_CLAIMED = 'POST_ALREADY_CLAIMED',

  //SubCategory
  SUBCATEGORY_NOT_FOUND = 'SUBCATEGORY_NOT_FOUND',

  //Location
  LOCATION_NOT_FOUND = 'LOCATION_NOT_FOUND',

  //Claim
  DUPLICATE_CLAIM = 'DUPLICATE_CLAIM',
  CLAIM_NOT_FOUND = 'CLAIM_NOT_FOUND',
  OWNER_CLAIM_ALREADY_EXISTS = 'OWNER_CLAIM_ALREADY_EXISTS',

  //General
  VALIDATION_ERROR = 'VALIDATION_ERROR',
  INTERNAL_ERROR = 'INTERNAL_ERROR',
  SERVICE_UNAVAILABLE = 'SERVICE_UNAVAILABLE',
}